Fix transaction size influx query. Cleanup tests for clippy.
