+++ src/tangle/sources/mod.rs
@@ -2,16 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #[cfg(feature = "inx")]
-mod inx;
-mod memory;
-mod mongodb;
+pub(crate) mod inx;
+pub(crate) mod memory;
+pub(crate) mod mongodb;
 use std::ops::RangeBounds;
 
 use async_trait::async_trait;
 use futures::stream::BoxStream;
 
-#[cfg(test)]
-pub(crate) use self::memory::test::{get_in_memory_data, IN_MEM_MILESTONE};
 use super::ledger_updates::LedgerUpdateStore;
 use crate::types::{
     ledger::{BlockMetadata, MilestoneIndexTimestamp},
